refactor: 상태를 갖는 Cell 객체 사용하도록 기존 코드 변경
