+++ src/main/kotlin/minesweeper/model/Cells.kt
@@ -1,40 +1,41 @@
 package minesweeper.model
 
-data class Cells(
-    val cells: List<Cell>
-) : List<Cell> by cells {
+class Cells(
+    val cells: MutableList<Cell>
+) : MutableList<Cell> by cells {
 
     val mineCount
-        get(): Int = cells.count { it.isMine() }
+        get(): Int = cells.count { it.isMine }
 
     fun take(startIndex: Int, endIndex: Int): Cells = cells.subList(startIndex, endIndex).let(::Cells)
 
     fun mineCountIn(positions: Set<CellPosition>): Int =
-        cells.count { it.isMineIn(positions) }
+        cells.count { it.isMineAndIn(positions) }
 
     fun findClosedCellsIn(positions: Set<CellPosition>): List<Cell> =
         cells.filter { it.isClosedAndIn(positions) }
 
-    fun openAndSurroundingNonMineCells(position: Position, mineBoard: MineBoard) =
-        cells[position.position].openMeAndSurroundingNonMineCells(mineBoard)
+    fun isOpenedAt(position: Position): Boolean = cells[position.position].isOpened
+
+    fun openAt(position: Position): Cell = cells[position.position].open()
 
     companion object {
         fun of(positions: CellPositions, mineCellCount: Int): Cells {
             val cellPositions = positions.toList()
 
-            val mineCells = List(mineCellCount) {
-                Cell.mine(cellPositions[it])
+            val closedMines = List(mineCellCount) {
+                ClosedMine(cellPositions[it])
             }
 
-            val closeCellCount = positions.size - mineCellCount
-            val closeCells = List(closeCellCount) {
-                Cell.nonMine(cellPositions[mineCellCount + it])
+            val closedNonMineCount = positions.size - mineCellCount
+            val closedNonMineCells = List(closedNonMineCount) {
+                ClosedNonMine(cellPositions[mineCellCount + it])
             }
 
-            val sortedCells = mineCells.plus(closeCells)
+            val sortedCells = closedMines.plus(closedNonMineCells)
                 .sortedBy { it.position }
 
-            return Cells(sortedCells)
+            return Cells(sortedCells.toMutableList())
         }
     }
 }
