+++ src/main/kotlin/minesweeper/model/ClosedNonMine.kt
@@ -1,6 +1,4 @@
-package minesweeper.model2
-
-import minesweeper.model.CellPosition
+package minesweeper.model
 
 class ClosedNonMine(
     position: CellPosition,
