+++ src/main/kotlin/minesweeper/model/OpenedNonMine.kt
@@ -1,6 +1,4 @@
-package minesweeper.model2
-
-import minesweeper.model.CellPosition
+package minesweeper.model
 
 class OpenedNonMine(
     position: CellPosition,
